Remove WinAPI at Render

+++ luke_engine_source/Input.cpp
@@ -93,10 +93,10 @@ namespace luke {
 		UINT height = application.GetHeight();
 		mMousePosition.x = -1.0f;
 		mMousePosition.y = -1.0f;
-		if (mousePos.x > 0 && mousePos.x < width)
-			mMousePosition.x = mousePos.x;
-		if (mousePos.y > 0 && mousePos.y < height)
-			mMousePosition.y = mousePos.y;
+		if ((UINT)mousePos.x > 0 && (UINT)mousePos.x < width)
+			mMousePosition.x = (float)mousePos.x;
+		if ((UINT)mousePos.y > 0 && (UINT)mousePos.y < height)
+			mMousePosition.y = (float)mousePos.y;
 	}
 	void Input::clearKeys()
 	{
